[CL-2106] Fix issue with new pages
